Spotlight & Color Traversal Animation In-Progress...
On Mouse Move & Color Traversal Animation only working on the first service's button, but not the Spotlight Effect. Alternatively, the rest of the service's buttons have the Spotlight Effect, but only in the top left corner as On Mouse Move isn't working. Additionally, those other service buttons do not have the Color Traversal Animation working either.